% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustCCx.R
\name{AdjustCCx}
\alias{AdjustCCx}
\title{Adjust CCx value for changes in CGC due to water stress during the growing season}
\usage{
AdjustCCx(CCprev, CCo, CCx, CGC, CDC, dt, tSum, Crop)
}
\arguments{
\item{CCprev}{Prev Cannopy cover}

\item{CCo}{initial canopy cover at the time of 90\% crop emergence}

\item{CCx}{Maximum canopy cover}

\item{CGC}{Canopy growth coefficient}

\item{CDC}{Canopy decline coefficient}

\item{dt}{Delta time}

\item{tSum}{tSum}

\item{Crop}{Parameters for a given crop}
}
\value{
with \code{CCxAdj} for a n time-step.
}
\description{
Adjust CCx value for changes in CGC due to water stress during the growing season
}
\examples{
AdjustCCx(CCprev, CCo, CCx, CGC, CDC, dt, tSum, Crop)
}
