% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AerationStress.R
\name{AerationStress}
\alias{AerationStress}
\title{Calculate aeration stress coefficient}
\usage{
AerationStress(Crop, InitCond, thRZ)
}
\arguments{
\item{Crop}{Parameters for a given crop}

\item{InitCond}{Crop setting initial structure}

\item{thRZ}{aeration stress (root zone)}
}
\value{
list with \code{NewCond} and \code{Ksa} aeration stress coefficient for a n time-step.
}
\description{
Calculate aeration stress coefficient
}
\examples{
AerationStress(Crop, InitCond, thRZ)
}
