% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aqua_library.R
\name{as_date_list}
\alias{as_date_list}
\title{It's datavec in Matlab. Converts data in numeric format to date in vector
format i.e. yyyy, m, d}
\usage{
as_date_list(sdat, origin_value = "0000-01-01")
}
\description{
It's datavec in Matlab. Converts data in numeric format to date in vector
format i.e. yyyy, m, d
}
