% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aqua_library.R
\name{as_datenum}
\alias{as_datenum}
\title{Number of days from sdat to origin}
\usage{
as_datenum(sdat, o = "0000-01-01", f = "\%Y-\%m-\%d")
}
\arguments{
\item{sdat}{date in format "\%Y-\%m-\%d" as default}

\item{o}{origin}

\item{f}{format}
}
\description{
Number of days from sdat to origin
}
