% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aqua_library.R
\name{as_datenum_string}
\alias{as_datenum_string}
\title{Number of days from sdat to origin}
\usage{
as_datenum_string(x)
}
\arguments{
\item{x}{date where x should be in the format: x['Year'], x['Month'], x['Day']}
}
\description{
Number of days from sdat to origin
}
