% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aqua_library.R
\name{calcAVP}
\alias{calcAVP}
\title{calculate Water vapour pressure (Kpa) according to FAO}
\usage{
calcAVP(x)
}
\arguments{
\item{x}{list with three parameters x[[1]] = RH, x[[2]] = TMAX, x[[3]] = TMIN}

\item{AVP}{actual vapour pressure}
}
\description{
calculate Water vapour pressure (Kpa) according to FAO
}
\examples{
RH = 57.8
TMAX = 22.96
TMIN = 9.23
calcAVP(RH, TMAX, TMIN)
}
