% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateHIGC.R
\name{CalculateHIGC}
\alias{CalculateHIGC}
\title{Calculate harvest index growth coefficient.}
\usage{
CalculateHIGC(Crop)
}
\arguments{
\item{Crop}{list}
}
\value{
list with \code{HIGC} and \code{tHI}.
}
\description{
Calculate harvest index growth coefficient.
}
\examples{
CalculateHIGC(Crop)
}
