% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateHILinear.R
\name{CalculateHILinear}
\alias{CalculateHILinear}
\title{Calculate time to switch to linear harvest index build-up, and 
associated linear rate of build-up. Only for fruit/grain crops.}
\usage{
CalculateHILinear(Crop)
}
\arguments{
\item{Crop}{list}
}
\value{
list with \code{tSwitch} and \code{dHILin}.
}
\description{
Calculate time to switch to linear harvest index build-up, and 
associated linear rate of build-up. Only for fruit/grain crops.
}
\examples{
CalculateHILinear(Crop)
}
