% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanopyCover.R
\name{CanopyCover}
\alias{CanopyCover}
\title{Simulate canopy growth/decline}
\usage{
CanopyCover(Crop, Soil, InitCond, GDD, Et0, GrowingSeason)
}
\arguments{
\item{Crop}{Parameters for a given crop}

\item{Soil}{properties of soil}

\item{InitCond}{Crop setting initial structure}

\item{GDD}{Growing degree days}

\item{Et0}{Evapotranspiration}

\item{GrowingSeason}{crop developmental stage}
}
\value{
\code{NewCond} for a n time-step.
}
\description{
Simulate canopy growth/decline
}
\examples{
CanopyCover(Crop, Soil, InitCond, GDD, Et0, GrowingSeason)
}
