% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CapillaryRise.R
\name{CapillaryRise}
\alias{CapillaryRise}
\title{Calculate capillary rise from a shallow groundwater table}
\usage{
CapillaryRise(Soil, Groundwater, InitCond, FluxOut)
}
\arguments{
\item{Soil}{properties of soil}

\item{Groundwater}{ground water table}

\item{InitCond}{Crop setting initial structure}

\item{FluxOut}{Flux}
}
\value{
list with \code{NewCond} and \code{CrTot} for a n time-step.
}
\description{
Calculate capillary rise from a shallow groundwater table
}
\examples{
CapillaryRise(Soil, Groundwater, InitCond, FluxOut)
}
