% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCDevelopment.R
\name{CCDevelopment}
\alias{CCDevelopment}
\title{Calculate canopy cover development by end of the current
simulation day}
\usage{
CCDevelopment(CCo, CCx, CGC, CDC, dt, Mode)
}
\arguments{
\item{CCo}{CCDevelopment}

\item{CCx}{Maximum canopy cover}

\item{CGC}{Canopy growth coefficient}

\item{CDC}{Canopy decline coefficient}

\item{dt}{Canopy approaching maximum size}

\item{Mode}{crop calendar mode}
}
\value{
\code{CC} canopy for a n time-step.
}
\description{
Calculate canopy cover development by end of the current
simulation day
}
\examples{
CCDevelopment(CCo, CCx, CGC, CDC, dt, Mode)
}
