% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRequiredTime.R
\name{CCRequiredTime}
\alias{CCRequiredTime}
\title{Function to find time required to reach CC at end of previous day, given
current CGC or CDC}
\usage{
CCRequiredTime(CCprev, CCo, CCx, CGC, CDC, dt, tSum, Mode)
}
\arguments{
\item{CCprev}{Prev Cannopy cover}

\item{CCo}{Initial canopy cover at the time of 90\% crop emergence}

\item{CCx}{Maximum canopy cover}

\item{CGC}{Canopy growth coefficient}

\item{CDC}{Canopy decline coefficient}

\item{dt}{delta time}

\item{tSum}{tSum}

\item{Mode}{Stage}
}
\value{
\code{tReq} eequired time for a n time-step.
}
\description{
Function to find time required to reach CC at end of previous day, given
current CGC or CDC
}
\examples{
CCRequiredTime(CCprev, CCo, CCx, CGC, CDC, dt, tSum, Mode)
}
