% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aqua_library.R
\name{ChangeDateFormat}
\alias{ChangeDateFormat}
\title{change date format}
\usage{
ChangeDateFormat(x)
}
\arguments{
\item{x}{date, where year = x[[1]], DOY = x[[2]]}
}
\value{
d
}
\description{
change date format
}
\examples{
year = '2011'
DOY = 1
ChangeDateFormat(year, DOY)

}
