% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckModelTermination.R
\name{CheckModelTermination}
\alias{CheckModelTermination}
\title{Function to check and declare model termination}
\usage{
CheckModelTermination(ClockStruct, InitialiseStruct)
}
\arguments{
\item{ClockStruct}{the Clock list}

\item{InitialiseStruct}{the crop initial conditions in list format}
}
\value{
\code{ClockStruct}.
}
\description{
Function to check and declare model termination
}
\examples{
CheckModelTermination(ClockStruct, InitialiseStruct)
}
