% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCropCalendar.R
\name{ComputeCropCalendar}
\alias{ComputeCropCalendar}
\title{Compute additional parameters needed to define crop}
\usage{
ComputeCropCalendar(Crop, CropName, CropChoices, Weather, ClockStruct)
}
\arguments{
\item{Crop}{list}

\item{CropName}{list with crops names}

\item{CropChoices}{crops to be analysed}

\item{Weather}{dataset with weather data}

\item{ClockStruct}{crop calendar}
}
\value{
\code{Crop}.
}
\description{
Compute additional parameters needed to define crop
}
\examples{
ComputeCropCalendar(Crop, CropName, CropChoices, Weather, ClockStruct)
}
