% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeVariables.R
\name{ComputeVariables}
\alias{ComputeVariables}
\title{Compute additional variables needed to run AquaCrop}
\usage{
ComputeVariables(ParamStruct, Weather, ClockStruct, GwStruct, CropChoices,
  FileLocation)
}
\arguments{
\item{ParamStruct}{Crop Structure}

\item{Weather}{Weather data}

\item{ClockStruct}{Crop calendar}

\item{GwStruct}{Ground water table}

\item{CropChoices}{Crops to be analysed}

\item{FileLocation}{list with file locations}
}
\value{
\code{ParamStruct}.
}
\description{
Compute additional variables needed to run AquaCrop
}
\examples{
ComputeVariables(ParamStruct, Weather, ClockStruct, GwStruct, CropChoices, FileLocation)

}
