% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aqua_library.R
\name{convert_list_2numeric}
\alias{convert_list_2numeric}
\title{convert list values to numeric}
\usage{
convert_list_2numeric(slist, lnames = NULL)
}
\arguments{
\item{slist}{list}

\item{lnames}{field names to transform
FIXME: do it for subsets}
}
\description{
convert list values to numeric
}
