% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CropParameters.R
\name{CropParameters}
\alias{CropParameters}
\title{Crop Parameters to be provided in CropFile.}
\usage{
CropParameters()
}
\description{
Crop Parameters to be provided in CropFile.
}
\references{
\insertAllCited{}
}
