% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drainage.R
\name{Drainage}
\alias{Drainage}
\title{Redistribute stored soil water}
\usage{
Drainage(Soil, InitCond)
}
\arguments{
\item{Soil}{structure characteristics}

\item{InitCond}{Crop setting initial structure}
}
\value{
\code{NewCond}, \code{DeepPerc} and \code{FluxOut} for n time-step
}
\description{
Redistribute stored soil water
}
\examples{
Drainage(Soil, InitCond)

}
