% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapLayerWaterContent.R
\name{EvapLayerWaterContent}
\alias{EvapLayerWaterContent}
\title{Get water contents in the evaporation layer}
\usage{
EvapLayerWaterContent(InitCond, Soil, Wevap)
}
\arguments{
\item{InitCond}{Crop setting initial structure}

\item{Soil}{properties of soil}

\item{wevap}{watwe evaporation}
}
\value{
\code{Wevap} for a n time-step.
}
\description{
Get water contents in the evaporation layer
}
\examples{
EvapLayerWaterContent(InitCond, Soil, Wevap)
}
