% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aqua_library.R
\name{export_as_xml}
\alias{export_as_xml}
\title{export dataframe as xml}
\usage{
export_as_xml(sdata, filename)
}
\arguments{
\item{sdata}{dataset}

\item{filename}{name of file}
}
\description{
export dataframe as xml
}
