% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractWeatherData.R
\name{ExtractWeatherData}
\alias{ExtractWeatherData}
\title{Extract weather data for current time step.}
\usage{
ExtractWeatherData(InitialiseStruct)
}
\arguments{
\item{InitialiseStruct}{Crop setting initial structure}

\item{ClockStruct}{crop calendar}
}
\value{
list with \code{Weather} for n time-step
}
\description{
Extract weather data for current time step.
}
\examples{
ExtractWeatherData(ClockStruct, InitialiseStruct)
}
