% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germination.R
\name{Germination}
\alias{Germination}
\title{Check if crop has germinated}
\usage{
Germination(InitCond, Soil, Crop, GDD, GrowingSeason)
}
\arguments{
\item{InitCond}{Crop setting initial structure}

\item{Soil}{properties of soil}

\item{Crop}{Parameters for a given crop}

\item{GDD}{Growing degree days}

\item{GrowingSeason}{crop developmental stage}
}
\value{
\code{NewCond}  for a n time-step.
}
\description{
Check if crop has germinated
}
\examples{
Germination(InitCond, Soil, Crop, GDD, GrowingSeason)
}
