% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aqua_library.R
\name{get_day}
\alias{get_day}
\title{get parameter day from date string}
\usage{
get_day(x)
}
\arguments{
\item{x}{date, where year = x[[1]], DOY = x[[2]]}
}
\value{
day of the month
}
\description{
get parameter day from date string
}
