% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aqua_library.R
\name{get_month}
\alias{get_month}
\title{get parameter month from date string}
\usage{
get_month(x)
}
\arguments{
\item{x}{date, where year = x[[1]], DOY = x[[2]]}
}
\value{
month of the year
}
\description{
get parameter month from date string
}
