% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aqua_library.R
\name{get_year}
\alias{get_year}
\title{get parameter year from date string}
\usage{
get_year(x)
}
\arguments{
\item{x}{date, where year = x[[1]], DOY = x[[2]]}
}
\value{
year
}
\description{
get parameter year from date string
}
