% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroundwaterInflow.R
\name{GroundwaterInflow}
\alias{GroundwaterInflow}
\title{Calculate capillary rise in the presence of a shallow
groundwater table}
\usage{
GroundwaterInflow(Soil, InitCond)
}
\arguments{
\item{Soil}{properties of soil}

\item{InitCond}{Crop setting initial structure}
}
\value{
list with \code{NewCond} and \code{GwIn} groundwater inflow for a n time-step.
}
\description{
Calculate capillary rise in the presence of a shallow
groundwater table
}
\examples{
GroundwaterInflow(Soil, InitCond)
}
