% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowingDegreeDay.R
\name{GrowingDegreeDay}
\alias{GrowingDegreeDay}
\title{Calculate number of growing degree days on current day.}
\usage{
GrowingDegreeDay(Crop, InitCond, Tmax, Tmin)
}
\arguments{
\item{Crop}{structure}

\item{InitCond}{initial crop settings}

\item{Tmax}{max temperature for n time-step}

\item{Tmin}{min temperature for n time-step}
}
\value{
list with \code{NewCond} and \code{GDD} for n time-step
}
\description{
Calculate number of growing degree days on current day.
}
\examples{
GrowingDegreeDay(Crop, InitCond, Tmax, Tmin)
}
