% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthStage.R
\name{GrowthStage}
\alias{GrowthStage}
\title{Calculate number of growing degree days on current day}
\usage{
GrowthStage(Crop, InitCond, GrowingSeason)
}
\arguments{
\item{Crop}{Parameters for a given crop}

\item{InitCond}{Crop setting initial structure}

\item{GrowingSeason}{crop developmental stage}
}
\value{
\code{NewCond}  for a n time-step.
}
\description{
Calculate number of growing degree days on current day
}
\examples{
GrowthStage(Crop, InitCond, GrowingSeason)
}
