% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestIndex.R
\name{HarvestIndex}
\alias{HarvestIndex}
\title{# Function to simulate build up of harvest index}
\usage{
HarvestIndex(Soil, Crop, InitCond, Et0, Tmax, Tmin, GDD, GrowingSeason)
}
\arguments{
\item{Soil}{properties of soil}

\item{Crop}{Parameters for a given crop}

\item{InitCond}{Crop setting initial structure}

\item{Et0}{Evapotranspiration}

\item{Tmax}{max temp for n time-step}

\item{Tmin}{min temp for n time-step}

\item{GDD}{Growing degree days}

\item{GrowingSeason}{crop developmental stage}
}
\value{
\code{NewCond} for a n time-step.
}
\description{
# Function to simulate build up of harvest index
}
\examples{
HarvestIndex(Soil, Crop, InitCond, Et0, Tmax, Tmin, GDD, GrowingSeason)
}
