% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIadjPollination.R
\name{HIadjPollination}
\alias{HIadjPollination}
\title{Calculate adjustment to harvest index for failure of
pollination due to water or temperature stress}
\usage{
HIadjPollination(InitCond, Crop, Ksw, Kst, HIt)
}
\arguments{
\item{InitCond}{Crop setting initial structure}

\item{Crop}{Parameters for a given crop}

\item{Ksw}{Water stress}

\item{Kst}{Temperature stress}

\item{HIt}{Harvest index on current day}
}
\value{
with \code{NewCond} for a n time-step.
}
\description{
Calculate adjustment to harvest index for failure of
pollination due to water or temperature stress
}
\examples{
HIadjPollination(InitCond, Crop, Ksw, Kst, HIt)
}
