% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIadjPostAnthesis.R
\name{HIadjPostAnthesis}
\alias{HIadjPostAnthesis}
\title{Function to calculate adjustment to harvest index for post-anthesis water
stress}
\usage{
HIadjPostAnthesis(InitCond, Crop, Ksw)
}
\arguments{
\item{InitCond}{Crop setting initial structure}

\item{Crop}{Parameters for a given crop}

\item{Ksw}{}
}
\value{
\code{NewCond} for a n time-step.
}
\description{
Function to calculate adjustment to harvest index for post-anthesis water
stress
}
\examples{
HIadjPostAnthesis(InitCond, Crop, Ksw)
}
