% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIadjPreAnthesis.R
\name{HIadjPreAnthesis}
\alias{HIadjPreAnthesis}
\title{Function to calculate adjustment to harvest index for pre-anthesis water
stress}
\usage{
HIadjPreAnthesis(InitCond, Crop)
}
\arguments{
\item{InitCond}{Crop setting initial structure}

\item{Crop}{Parameters for a given crop}
}
\value{
\code{NewCond} for a n time-step.
}
\description{
Function to calculate adjustment to harvest index for pre-anthesis water
stress
}
\examples{
HIadjPreAnthesis(InitCond, Crop)
}
