% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIrefCurrentDay.R
\name{HIrefCurrentDay}
\alias{HIrefCurrentDay}
\title{Calculate reference (no adjustment for stress effects)
harvest index on current day}
\usage{
HIrefCurrentDay(InitCond, Crop, GrowingSeason)
}
\arguments{
\item{InitCond}{Crop setting initial structure}

\item{Crop}{Parameters for a given crop}

\item{GrowingSeason}{crop developmental stage}
}
\value{
\code{NewCond} for a n time-step.
}
\description{
Calculate reference (no adjustment for stress effects)
harvest index on current day
}
\examples{
HIrefCurrentDay(InitCond, Crop, GrowingSeason)
}
