% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Initialise.R
\name{Initialise}
\alias{Initialise}
\title{Function to initialise AquaCropR}
\usage{
Initialise(FileLocation)
}
\arguments{
\item{filename}{An xml file with locations}
}
\value{
list with \code{FileLocation}.
}
\description{
Function to initialise AquaCropR
}
\examples{
ReadFileLocations('dummy.xml')
}
