% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformSimulation.R
\name{PerformSimulation}
\alias{PerformSimulation}
\title{Perform simulation}
\usage{
PerformSimulation(InitialiseStruct)
}
\arguments{
\item{InitialiseStruct}{Crop setting initial structure}
}
\value{
list with \code{Outputs} results.
}
\description{
Perform simulation
}
\examples{
PerfomSimulation(InitialiseStruct)
}
