% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreIrrigation.R
\name{PreIrrigation}
\alias{PreIrrigation}
\title{Calculate pre-irrigation when in net irrigation mode}
\usage{
PreIrrigation(Soil, Crop, IrrMngt, InitCond)
}
\arguments{
\item{Soil}{structure characteristics}

\item{Crop}{crop settings}

\item{IrrMngt}{irigation management settings}

\item{InitCond}{Crop setting initial structure}
}
\value{
list with \code{NewCond} and \code{PreIrr} for n time-step
}
\description{
Calculate pre-irrigation when in net irrigation mode
}
\examples{
PreIrrigation((Soil, Crop, IrrMngt, InitCond)
}
