% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RainfallPartition.R
\name{RainfallPartition}
\alias{RainfallPartition}
\title{Partition rainfall into surface runoff and infiltration}
\usage{
RainfallPartition(P, Soil, FieldMngt, InitCond)
}
\arguments{
\item{P}{precipitation}

\item{Soil}{structure characteristics}

\item{InitCond}{Crop setting initial structure}

\item{IrrMngt}{Irrigation management}
}
\value{
list with \code{NewCond}, \code{Runoff} and \code{Infl} for n time-step
}
\description{
Partition rainfall into surface runoff and infiltration
}
\examples{
RainfallPartition(P, Soil, FieldMngt, InitCond)
}
