% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadClockParameters.R
\name{ReadClockParameters}
\alias{ReadClockParameters}
\title{Read input files and initialise model clock parameters}
\usage{
ReadClockParameters(FileLocation)
}
\arguments{
\item{FileLocation}{list with file locations}
}
\value{
\code{ClockStruct} crop calendar.
}
\description{
Read input files and initialise model clock parameters
}
\examples{
ReadClockParameters(FileLocation)
}
