% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFieldManagement.R
\name{ReadFieldManagement}
\alias{ReadFieldManagement}
\title{Read input files and initialise model clock parameters}
\usage{
ReadFieldManagement(FileLocation)
}
\arguments{
\item{FileLocation}{list with file locations}
}
\value{
\code{FieldManagement} list.
}
\description{
Read input files and initialise model clock parameters
}
\examples{
ReadFieldManagement(FileLocation)
}
