% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGroundwaterTable.R
\name{ReadGroundwaterTable}
\alias{ReadGroundwaterTable}
\title{Read input file and initialise groundwater table parameters}
\usage{
ReadGroundwaterTable(FileLocation, ClockStruct)
}
\arguments{
\item{FileLocation}{list with file locations}

\item{ClockStruct}{crop calendar}
}
\value{
\code{GwStruct} water table.
}
\description{
Read input file and initialise groundwater table parameters
}
\examples{
ReadGroundwaterTable(FileLocation, ClockStruct)
FIXME there is abug here
}
