% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadIrrigationManagement.R
\name{ReadIrrigationManagement}
\alias{ReadIrrigationManagement}
\title{Compute additional variables needed to run AOS}
\usage{
ReadIrrigationManagement(ParamStruct, FileLocation, ClockStruct)
}
\arguments{
\item{ParamStruct}{Crop parameters structure}

\item{FileLocation}{list with file locations}

\item{ClockStruct}{crop calendar}
}
\value{
\code{ParamStruct}.
}
\description{
Compute additional variables needed to run AOS
}
\examples{
ReadIrrigationManagement(ParamStruct, FileLocation, ClockStruct)
}
