% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadModelInitialConditions.R
\name{ReadModelInitialConditions}
\alias{ReadModelInitialConditions}
\title{Set up initial model conditions}
\usage{
ReadModelInitialConditions(ParamStruct, GwStruct, FieldMngtStruct,
  CropChoices, FileLocation, ClockStruct)
}
\arguments{
\item{ParamStruct}{Crop details}

\item{GwStruct}{Water table details}

\item{FieldMngtStruct}{field management structure}

\item{CropChoices}{crops whose performance is to be modelled}

\item{FileLocation}{file locations}

\item{ClockStruct}{crop calendar structure}
}
\value{
\code{InitCondStruct} with intial conditions.
}
\description{
Set up initial model conditions
}
\examples{
ReadModelInitialConditions(ParamStruct, GwStruct, FieldMngtStruct, CropChoices, FileLocation,
ClockStruct)
}
