% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadModelParameters.R
\name{ReadModelParameters}
\alias{ReadModelParameters}
\title{Read input files and initialise soil and crop parameters}
\usage{
ReadModelParameters(FileLocation, ClockStruct)
}
\arguments{
\item{FileLocation}{list with file locations}

\item{ClockStruct}{crop calendar}
}
\value{
list with uodated \code{ClockStruct} \code{ParamStruct} and \code{CropChoices}.
}
\description{
Read input files and initialise soil and crop parameters
}
\examples{
ReadModelParameters(FileLocation, ClockStruct)
}
