% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWeatherInputs.R
\name{ReadWeatherInputs}
\alias{ReadWeatherInputs}
\title{Read and process input weather time-series}
\usage{
ReadWeatherInputs(FileLocation, ClockStruct)
}
\arguments{
\item{FileLocation}{file locations}

\item{ClockStruct}{list time}
}
\value{
list with \code{WeatherDB}.
}
\description{
Read and process input weather time-series
}
\examples{
ReadWeatherInputs(FileLocation, ClockStruct)
}
