% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResetInitialConditions.R
\name{ResetInitialConditions}
\alias{ResetInitialConditions}
\title{Reset initial model conditions for start of growing}
\usage{
ResetInitialConditions(InitialiseStruct, ClockStruct)
}
\arguments{
\item{InitialiseStruct}{Crop setting initial structure}

\item{ClockStruct}{crop calendar}
}
\value{
list with \code{InitialiseStruct} and \code{ClockStruct} for a n time-step
}
\description{
Reset initial model conditions for start of growing
}
\examples{
ResetInitialConditions(InitialiseStruct, ClockStruct)
}
