% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RootDevelopment.R
\name{RootDevelopment}
\alias{RootDevelopment}
\title{Calculate root zone expansion}
\usage{
RootDevelopment(Crop, Soil, GroundWater, InitCond, GDD, GrowingSeason)
}
\arguments{
\item{Crop}{Parameters for a given crop}

\item{Soil}{properties of soil}

\item{InitCond}{Crop setting initial structure}

\item{GDD}{Growing degree days}

\item{GrowingSeason}{crop developmental stage}

\item{Groundwater}{ground water table}
}
\value{
\code{NewCond} for a n time-step.
}
\description{
Calculate root zone expansion
}
\examples{
RootDevelopment(Crop, Soil, GroundWater, InitCond, GDD, GrowingSeason)
}
