% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RootZoneWater.R
\name{RootZoneWater}
\alias{RootZoneWater}
\title{Calculate actual and total available water in the root}
\usage{
RootZoneWater(Soil, Crop, InitCond)
}
\arguments{
\item{Soil}{properties of soil}

\item{Crop}{parameters for a given crop}

\item{InitCond}{Crop setting initial structure}
}
\value{
list with \code{Wr}, \code{Dr},  \code{TAW} and \code{thRZ} for a n time-step.
}
\description{
Calculate actual and total available water in the root
}
\examples{
RootZoneWater(Soil, Crop, InitCond)
}
