% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilEvaporation.R
\name{SoilEvaporation}
\alias{SoilEvaporation}
\title{Calculate daily soil evaporation in AOS}
\usage{
SoilEvaporation(ClockStruct, Soil, Crop, IrrMngt, FieldMngt, InitCond, Et0,
  Infl, Rain, Irr, GrowingSeason)
}
\arguments{
\item{ClockStruct}{crop calendar}

\item{Soil}{properties of soil}

\item{Crop}{Parameters for a given crop}

\item{IrrMngt}{Irrigation management}

\item{FieldMngt}{Field management}

\item{InitCond}{Crop setting initial structure}

\item{Et0}{Evapotranspiration}

\item{Infl}{Infiltration}

\item{Irr}{Irrigation}

\item{GrowingSeason}{crop developmental stage}

\item{P}{precipitation}
}
\value{
list with \code{NewCond}, \code{EsAct} actual evaporation and \code{EsPot} 
potential soil evaporation for a n time-step.
}
\description{
Calculate daily soil evaporation in AOS
}
\examples{
SoilEvaporation(ClockStruct, Soil, Crop, IrrMngt, FieldMngt,InitCond,Et0,Infl,Rain,Irr,GrowingSeason)
}
