% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilHydraulicProperties.R
\name{SoilHydraulicProperties}
\alias{SoilHydraulicProperties}
\title{Calculate soil hydraulic properties, given textural inputs.
Calculations use pedotransfer function equations described in Saxton and
Rawls (2006)}
\usage{
SoilHydraulicProperties(Soil)
}
\arguments{
\item{Soil}{properties of soil
#' @return list with \code{thdry}, \code{thwp}, \code{hfc}, \code{ths} and \code{ksa}
for a n time-step.}
}
\description{
Calculate soil hydraulic properties, given textural inputs.
Calculations use pedotransfer function equations described in Saxton and
Rawls (2006)
}
\examples{
SoilHydraulicProperties(Soil)
}
