% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solution.R
\name{Solution}
\alias{Solution}
\title{Run AquaCrop for a n time step}
\usage{
Solution(Weather, InitialiseStruct)
}
\arguments{
\item{Weather}{weather parameters or n time-step}

\item{InitialiseStruct}{Crop setting initial structure}
}
\value{
list with \code{NewCond}, \code{Outputs} and \code{ClockStruct} for n time-step
}
\description{
Run AquaCrop for a n time step
}
\examples{
Solution(Weather, ClockStruct, InitialiseStruct)
}
