% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TemperatureStress.R
\name{TemperatureStress}
\alias{TemperatureStress}
\title{Calculate temperature stress coefficients}
\usage{
TemperatureStress(Crop, Tmax, Tmin, GDD)
}
\arguments{
\item{Crop}{Parameters for a given crop}

\item{Tmax}{max temp for n time-step}

\item{Tmin}{min temp for n time-step}

\item{GDD}{Growing degree days}
}
\value{
\code{Kst} temperature stress a n time-step.
}
\description{
Calculate temperature stress coefficients
}
\examples{
TemperatureStress(Crop, Tmax, Tmin, GDD)
}
