% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transpiration.R
\name{Transpiration}
\alias{Transpiration}
\title{Calculate crop transpiration on current day}
\usage{
Transpiration(Soil, Crop, IrrMngt, InitCond, Et0, CO2, GrowingSeason)
}
\arguments{
\item{Soil}{properties of soil}

\item{Crop}{Parameters for a given crop}

\item{IrrMngt}{Irrigation management}

\item{InitCond}{Crop setting initial structure}

\item{Et0}{Evapotranspiration}

\item{CO2}{Infiltration}

\item{GrowingSeason}{crop developmental stage}
}
\value{
list with \code{NewCond}, \code{TrAct} actual transpiration, \code{TrPot_NS} potential transpiration rate no water stress, 
\code{TrPot0} potential transpiration rate, \code{NewCond} and \code{IrrNet} Initialise net irrigation, for a n time-step.
}
\description{
Calculate crop transpiration on current day
}
\examples{
Transpiration(Soil, Crop, IrrMngt, InitCond, Et0, CO2, GrowingSeason)
}
