% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateTime.R
\name{UpdateTime}
\alias{UpdateTime}
\title{Update current time in model}
\usage{
UpdateTime(ClockStruct, InitialiseStruct)
}
\arguments{
\item{ClockStruct}{crop calendar}

\item{InitialiseStruct}{Crop setting initial structure}
}
\value{
list with \code{InitialiseStruct} and \code{ClockStruct} for a n time-step
}
\description{
Update current time in model
}
\examples{
UpdateTime(ClockStruct, InitialiseStruct)
}
