% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterStress.R
\name{WaterStress}
\alias{WaterStress}
\title{Calculate relative root zone water depletion for each stress type}
\usage{
WaterStress(Crop, InitCond, Dr, TAW, Et0, beta)
}
\arguments{
\item{Crop}{Parameters for a given crop}

\item{InitCond}{Crop setting initial structure}

\item{Dr}{Root zone water content at air dry}

\item{TAW}{Total available water}

\item{Et0}{Evapotranspiration}

\item{beta}{beta}
}
\value{
water stress list with \code{Exp}, \code{Sto}, \code{Sen}, \code{Pol} 
and \code{StoLin} for a n time-step.
}
\description{
Calculate relative root zone water depletion for each stress type
}
\examples{
WaterStress(Crop,InitCond,Dr,TAW,Et0, beta)
}
